working on preventing dupes in pokesearch
